fix: tighten time-decay exit to 1.5h/0.4% (was 3h/0.25%)
30-day trade data showed TP-hit trades resolve in a median of 46min
(max 99min), but MACD positions were drifting 1.6-5.7h unresolved
until the daily 15:45 EOD flatten cut them for a small loss (24
trades, net -$2.96). The old 3h threshold let dead capital sit far
past when a real winner would have shown itself.

Also switches getFlatPositionHours() from int to double — the old
int parser would silently reject a fractional value like "1.5" and
fall back to a hardcoded default of 2h, defeating the config change.

+++ trading-backend/src/main/java/com/trading/config/Config.java
@@ -1461,8 +1461,8 @@ public class Config {
         return getBooleanProperty("TIME_DECAY_EXITS", true);
     }
     
-    public int getFlatPositionHours() {
-        return getIntProperty("FLAT_POSITION_HOURS", 2);
+    public double getFlatPositionHours() {
+        return getDoubleProperty("FLAT_POSITION_HOURS", 2);
     }
     
     public double getFlatPositionThreshold() {
